API: cleanup of unit-tests
